improvement: included the reusable button #24
